exhuastiveness: wip move everything into `PatLowerCtx`
