+++ compiler/hash-typecheck/src/ops/reader.rs
@@ -1,13 +1,17 @@
 //! Contains helpers to read various things stored in [crate::storage] with
 //! ease.
 
-use crate::{storage::{
-    primitives::{
-        Args, ArgsId, ModDef, ModDefId, NominalDef, NominalDefId, Params, ParamsId, Pat, PatArgs,
-        PatArgsId, PatId, Scope, ScopeId, Term, TermId, TrtDef, TrtDefId, DeconstructedPatFieldsId,
+use crate::{
+    exhaustiveness::structures::DeconstructedPat,
+    storage::{
+        primitives::{
+            Args, ArgsId, DeconstructedPatId, ModDef, ModDefId, NominalDef, NominalDefId, Params,
+            ParamsId, Pat, PatArgs, PatArgsId, PatId, Scope, ScopeId, Term, TermId, TrtDef,
+            TrtDefId,
+        },
+        GlobalStorage,
     },
-    GlobalStorage,
-}, exhaustiveness::deconstruct::DeconstructedPat};
+};
 
 /// Helper to read various primitive constructions (from
 /// [crate::storage::primitives]).
@@ -67,8 +71,8 @@ impl<'gs> PrimitiveReader<'gs> {
     }
 
     /// Get the [Fields] with the given [DeconstructedPatFieldsId].
-    pub fn get_deconstructed_pat_fields(&self, id: DeconstructedPatFieldsId) -> &Vec<DeconstructedPat> {
-        self.gs.pat_fields_store.get(id)
+    pub fn get_deconstructed_pat_fields(&self, id: DeconstructedPatId) -> &DeconstructedPat {
+        self.gs.deconstructed_pat_store.get(id)
     }
 
     /// Get the trait definition with the given [TrtDefId].
